# Conflicts:
#	megamek/src/megamek/client/ui/SharedUtility.java
#	megamek/src/megamek/common/Aero.java
#	megamek/src/megamek/common/AlphaStrikeElement.java
#	megamek/src/megamek/common/BattleArmor.java
#	megamek/src/megamek/common/BattleForceElement.java
#	megamek/src/megamek/common/ConvFighter.java
#	megamek/src/megamek/common/Dropship.java
#	megamek/src/megamek/common/Entity.java
#	megamek/src/megamek/common/FixedWingSupport.java
#	megamek/src/megamek/common/Infantry.java
#	megamek/src/megamek/common/Jumpship.java
#	megamek/src/megamek/common/LandAirMech.java
#	megamek/src/megamek/common/Mech.java
#	megamek/src/megamek/common/Protomech.java
#	megamek/src/megamek/common/QuadVee.java
#	megamek/src/megamek/common/SmallCraft.java
#	megamek/src/megamek/common/Tank.java
#	megamek/src/megamek/common/UnitRole.java
#	megamek/src/megamek/common/VTOL.java
#	megamek/src/megamek/common/WeaponType.java
#	megamek/src/megamek/common/commandline/MegaMekCommandLineParser.java
#	megamek/src/megamek/common/weapons/autocannons/RACWeapon.java
#	megamek/src/megamek/common/weapons/battlearmor/CLBALBX.java
#	megamek/src/megamek/utilities/AlphaStrikeMassConvert.java

+++ megamek/unittests/megamek/common/UnitNameTrackerTest.java
@@ -1,22 +1,27 @@
 /*
- * Copyright (C) 2021 - The MegaMek Team
+ * Copyright (c) 2021-2022 - The MegaMek Team. All Rights Reserved.
  *
- *  This program is free software; you can redistribute it and/or modify it
- *  under the terms of the GNU General Public License as published by the Free
- *  Software Foundation; either version 2 of the License, or (at your option)
- *  any later version.
+ * This file is part of MegaMek.
  *
- *  This program is distributed in the hope that it will be useful, but
- *  WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- *  or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
- *  for more details.
+ * MegaMek is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MegaMek is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
  */
-
 package megamek.common;
 
-import static org.mockito.Mockito.*;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.*;
 
 public class UnitNameTrackerTest {
     @Test
@@ -194,7 +199,7 @@ public class UnitNameTrackerTest {
         Entity mockEntity = mock(Entity.class);
         when(mockEntity.getShortNameRaw()).thenReturn(shortNameRaw);
         doAnswer(inv -> {
-            int marker = (int) inv.getArgument(0);
+            int marker = inv.getArgument(0);
             when(mockEntity.getDuplicateMarker()).thenReturn(marker);
             return null;
         }).when(mockEntity).setDuplicateMarker(anyInt());
